# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Soumya Raikwar portfolio
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BlogPost {
  id: number;
  slug: string;
  title: string;
  excerpt: string;
  content: string;
  tags: string[];
  coverImage?: string | null;
  publishedAt: string;
  createdAt: string;
  updatedAt: string;
}

export interface BlogPostList {
  posts: BlogPost[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface CreateBlogPostBody {
  slug: string;
  title: string;
  excerpt: string;
  content: string;
  tags: string[];
  coverImage?: string | null;
  publishedAt?: string;
}

export interface UpdateBlogPostBody {
  title?: string;
  excerpt?: string;
  content?: string;
  tags?: string[];
  coverImage?: string | null;
  publishedAt?: string;
}

export interface SendMessageBody {
  name: string;
  email: string;
  subject: string;
  body: string;
}

export interface MessageSentResponse {
  success: boolean;
  message: string;
}

export interface TagList {
  tags: string[];
}

export type ListBlogPostsParams = {
  page?: number;
  limit?: number;
  tag?: string;
};