# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Soumya Raikwar portfolio
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all blog posts
 */
export const listBlogPostsQueryPageDefault = 1;
export const listBlogPostsQueryLimitDefault = 10;

export const ListBlogPostsQueryParams = zod.object({
  page: zod.coerce.number().default(listBlogPostsQueryPageDefault),
  limit: zod.coerce.number().default(listBlogPostsQueryLimitDefault),
  tag: zod.coerce.string().optional(),
});

export const ListBlogPostsResponse = zod.object({
  posts: zod.array(
    zod.object({
      id: zod.number(),
      slug: zod.string(),
      title: zod.string(),
      excerpt: zod.string(),
      content: zod.string(),
      tags: zod.array(zod.string()),
      coverImage: zod.string().nullish(),
      publishedAt: zod.coerce.date(),
      createdAt: zod.coerce.date(),
      updatedAt: zod.coerce.date(),
    }),
  ),
  total: zod.number(),
  page: zod.number(),
  limit: zod.number(),
  totalPages: zod.number(),
});

/**
 * @summary Create a blog post
 */
export const CreateBlogPostBody = zod.object({
  slug: zod.string(),
  title: zod.string(),
  excerpt: zod.string(),
  content: zod.string(),
  tags: zod.array(zod.string()),
  coverImage: zod.string().nullish(),
  publishedAt: zod.coerce.date().optional(),
});

/**
 * @summary Get a blog post by slug
 */
export const GetBlogPostParams = zod.object({
  slug: zod.coerce.string(),
});

export const GetBlogPostResponse = zod.object({
  id: zod.number(),
  slug: zod.string(),
  title: zod.string(),
  excerpt: zod.string(),
  content: zod.string(),
  tags: zod.array(zod.string()),
  coverImage: zod.string().nullish(),
  publishedAt: zod.coerce.date(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Update a blog post
 */
export const UpdateBlogPostParams = zod.object({
  slug: zod.coerce.string(),
});

export const UpdateBlogPostBody = zod.object({
  title: zod.string().optional(),
  excerpt: zod.string().optional(),
  content: zod.string().optional(),
  tags: zod.array(zod.string()).optional(),
  coverImage: zod.string().nullish(),
  publishedAt: zod.coerce.date().optional(),
});

export const UpdateBlogPostResponse = zod.object({
  id: zod.number(),
  slug: zod.string(),
  title: zod.string(),
  excerpt: zod.string(),
  content: zod.string(),
  tags: zod.array(zod.string()),
  coverImage: zod.string().nullish(),
  publishedAt: zod.coerce.date(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete a blog post
 */
export const DeleteBlogPostParams = zod.object({
  slug: zod.coerce.string(),
});

/**
 * @summary Get all unique tags used in blog posts
 */
export const GetBlogTagsResponse = zod.object({
  tags: zod.array(zod.string()),
});

/**
 * @summary Send a contact message
 */
export const SendMessageBody = zod.object({
  name: zod.string(),
  email: zod.string(),
  subject: zod.string(),
  body: zod.string(),
});

/**
 * @summary Get the 3 most recent blog posts
 */
export const GetRecentBlogPostsResponseItem = zod.object({
  id: zod.number(),
  slug: zod.string(),
  title: zod.string(),
  excerpt: zod.string(),
  content: zod.string(),
  tags: zod.array(zod.string()),
  coverImage: zod.string().nullish(),
  publishedAt: zod.coerce.date(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const GetRecentBlogPostsResponse = zod.array(
  GetRecentBlogPostsResponseItem,
);