# lib/api-zod/src/generated/types/sendMessageBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Soumya Raikwar portfolio
 * OpenAPI spec version: 0.1.0
 */

export interface SendMessageBody {
  name: string;
  email: string;
  subject: string;
  body: string;
}